refactor: relocate Statistics to the Client directory and create a shared Statistics page
