+++ apps/web/app/modules/Statistics/Client/components/ChartLegendBadge.tsx
