+++ apps/web/routes.ts
@@ -17,9 +17,7 @@ export const routes: (
       route("courses", "modules/Dashboard/Dashboard.page.tsx");
       route("course/:id", "modules/Courses/CourseView/CourseView.page.tsx");
       route("course/:courseId/lesson/:lessonId", "modules/Courses/Lesson/Lesson.page.tsx");
-      route("settings", "modules/Dashboard/Settings/Settings.page.tsx", {
-        index: true,
-      });
+      route("settings", "modules/Dashboard/Settings/Settings.page.tsx");
       route("tutors/:id", "modules/Tutor/Tutor.page.tsx");
     });
     route("admin", "modules/Admin/Admin.layout.tsx", () => {
